backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx==0.28.1
numpy==2.0.2
sentence-transformers==3.4.1
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20
